+++ example/unknown_dimension.cpp
@@ -26,7 +26,7 @@
 namespace {
 
 template<units::physical::Length D, units::physical::Time T>
-constexpr units::physical::Speed AUTO avg_speed(D d, T t)
+constexpr units::physical::Speed auto avg_speed(D d, T t)
 {
   return d / t;
 }
@@ -36,13 +36,13 @@ void example()
   using namespace units::physical;
   using namespace units::physical::si::literals;
 
-  Length AUTO d1 = 123q_m;
-  Time AUTO t1 = 10q_s;
-  Speed AUTO v1 = avg_speed(d1, t1);
+  Length auto d1 = 123q_m;
+  Time auto t1 = 10q_s;
+  Speed auto v1 = avg_speed(d1, t1);
 
   auto temp1 = v1 * 50q_m;  // produces intermediate unknown dimension with 'unknown_coherent_unit' as its 'coherent_unit'
-  Speed AUTO v2 = temp1 / 100q_m; // back to known dimensions again
-  Length AUTO d2 = v2 * 60q_s;
+  Speed auto v2 = temp1 / 100q_m; // back to known dimensions again
+  Length auto d2 = v2 * 60q_s;
 
   std::cout << "d1 = " << d1 << '\n';
   std::cout << "t1 = " << t1 << '\n';