+++ src/include/units/bits/common_quantity.h
@@ -76,15 +76,7 @@ using common_quantity_point = decltype(
 
 }  // namespace units
 
-#if COMP_MSVC || COMP_GCC >= 10
-
 namespace std {
-
-#else
-
-namespace concepts {
-
-#endif
 
 template<units::Quantity Q1, units::Quantity Q2>
   requires units::equivalent_dim<typename Q1::dimension, typename Q2::dimension>