+++ docs/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release notes
 
 - **0.6.0 WIP**
-  - Minimum compiler version supported changed to gcc-9.3
+  - gcc-9 is no longer supported (at least gcc-10 is required)
   - MSVC 16.7 support added
   - linear_algebra updated to 0.7.0/stable
   - fmt updated to 7.0.3